+++ src/allheaders.cpp
@@ -1,3 +1,0 @@
-#include <ptxall.hpp>
-
-int main () { return 0; }